requirements.txt
openai
requests
jsonschema
pydantic
pyyaml
tabulate